refactored & added deck draw mechanism
